fix: make skeleton manager service a render unit
