+++ packages/sheets-ui/src/services/sheet-skeleton-manager.service.ts
@@ -15,14 +15,13 @@
  */
 
 import type { Nullable, Workbook, Worksheet } from '@univerjs/core';
-import { IUniverInstanceService } from '@univerjs/core';
+import type { IRenderContext, IRenderController } from '@univerjs/engine-render';
 import { SpreadsheetSkeleton } from '@univerjs/engine-render';
 import type { IDisposable } from '@wendellhu/redi';
 import { Inject, Injector } from '@wendellhu/redi';
 import { BehaviorSubject } from 'rxjs';
 
 export interface ISheetSkeletonManagerParam {
-    unitId: string;
     sheetId: string;
     skeleton: SpreadsheetSkeleton;
     dirty: boolean;
@@ -30,7 +29,6 @@ export interface ISheetSkeletonManagerParam {
 }
 
 export interface ISheetSkeletonManagerSearch {
-    unitId: string;
     sheetId: string;
     commandId?: string; // WTF: why?
 }
@@ -46,9 +44,8 @@ export interface ISheetSkeletonManagerSearch {
  *
  * @todo RenderUnit - We should move this to RenderUnit as well after all dependents have been moved.
  */
-export class SheetSkeletonManagerService implements IDisposable {
+export class SheetSkeletonManagerService implements IDisposable, IRenderController {
     private _currentSkeleton: ISheetSkeletonManagerSearch = {
-        unitId: '',
         sheetId: '',
     };
 
@@ -64,7 +61,7 @@ export class SheetSkeletonManagerService implements IDisposable {
     readonly currentSkeletonBefore$ = this._currentSkeletonBefore$.asObservable();
 
     constructor(
-        @IUniverInstanceService private readonly _univerInstanceService: IUniverInstanceService,
+        private readonly _context: IRenderContext<Workbook>,
         @Inject(Injector) private readonly _injector: Injector
     ) {
         // empty
@@ -76,40 +73,21 @@ export class SheetSkeletonManagerService implements IDisposable {
         this._sheetSkeletonParam = [];
     }
 
+    getCurrentSkeleton(): SpreadsheetSkeleton {
+        return this.getCurrent()!.skeleton;
+    }
+
     /** @deprecated */
     getCurrent(): Nullable<ISheetSkeletonManagerParam> {
         return this._getSkeleton(this._currentSkeleton);
     }
 
     getUnitSkeleton(unitId: string, sheetId: string): Nullable<ISheetSkeletonManagerParam> {
-        return this._getSkeleton({ unitId, sheetId });
+        return this._getSkeleton({ sheetId });
     }
 
     setCurrent(searchParam: ISheetSkeletonManagerSearch): Nullable<ISheetSkeletonManagerParam> {
         this._setCurrent(searchParam);
-    }
-
-    removeSkeleton(searchParam: Pick<ISheetSkeletonManagerSearch, 'unitId'>) {
-        const index = this._sheetSkeletonParam.findIndex((param) => param.unitId === searchParam.unitId);
-        if (index !== -1) {
-            const skeletonParam = this._sheetSkeletonParam[index];
-            skeletonParam.skeleton.dispose();
-
-            this._sheetSkeletonParam.splice(index, 1);
-            this._currentSkeletonBefore$.next(null);
-            this._currentSkeleton$.next(null);
-        }
-    }
-
-    private _compareSearch(param1: Nullable<ISheetSkeletonManagerSearch>, param2: Nullable<ISheetSkeletonManagerSearch>) {
-        if (param1 == null || param2 == null) {
-            return false;
-        }
-
-        if (param1.commandId === param2.commandId && param1.sheetId === param2.sheetId && param1.unitId === param2.unitId) {
-            return true;
-        }
-        return false;
     }
 
     private _setCurrent(searchParam: ISheetSkeletonManagerSearch): Nullable<ISheetSkeletonManagerParam> {
@@ -117,19 +95,15 @@ export class SheetSkeletonManagerService implements IDisposable {
         if (param != null) {
             this._reCalculate(param);
         } else {
-            const { unitId, sheetId } = searchParam;
-
-            const workbook = this._univerInstanceService.getUniverSheetInstance(searchParam.unitId);
-
-            const worksheet = workbook?.getSheetBySheetId(searchParam.sheetId);
-            if (worksheet == null || workbook == null) {
+            const { sheetId } = searchParam;
+            const workbook = this._context.unit;
+            const worksheet = workbook.getSheetBySheetId(searchParam.sheetId);
+            if (worksheet == null) {
                 return;
             }
 
-            const skeleton = this._buildSkeleton(worksheet, workbook);
-
+            const skeleton = this._buildSkeleton(worksheet);
             this._sheetSkeletonParam.push({
-                unitId,
                 sheetId,
                 skeleton,
                 dirty: false,
@@ -177,15 +151,14 @@ export class SheetSkeletonManagerService implements IDisposable {
             return skeleton.skeleton;
         }
 
-        const workbook = this._univerInstanceService.getUniverSheetInstance(searchParam.unitId);
-        const worksheet = workbook?.getSheetBySheetId(searchParam.sheetId);
-        if (!worksheet || !workbook) {
+        const workbook = this._context.unit;
+        const worksheet = workbook.getSheetBySheetId(searchParam.sheetId);
+        if (!worksheet) {
             return;
         }
 
-        const newSkeleton = this._buildSkeleton(worksheet, workbook);
+        const newSkeleton = this._buildSkeleton(worksheet);
         this._sheetSkeletonParam.push({
-            unitId: searchParam.unitId,
             sheetId: searchParam.sheetId,
             skeleton: newSkeleton,
             dirty: false,
@@ -195,10 +168,7 @@ export class SheetSkeletonManagerService implements IDisposable {
     }
 
     private _getSkeleton(searchParm: ISheetSkeletonManagerSearch): Nullable<ISheetSkeletonManagerParam> {
-        const item = this._sheetSkeletonParam.find(
-            (param) => param.unitId === searchParm.unitId && param.sheetId === searchParm.sheetId
-        );
-
+        const item = this._sheetSkeletonParam.find((param) => param.sheetId === searchParm.sheetId);
         if (item != null) {
             item.commandId = searchParm.commandId;
         }
@@ -206,14 +176,14 @@ export class SheetSkeletonManagerService implements IDisposable {
         return item;
     }
 
-    private _buildSkeleton(worksheet: Worksheet, workbook: Workbook) {
+    private _buildSkeleton(worksheet: Worksheet) {
         const config = worksheet.getConfig();
         const spreadsheetSkeleton = this._injector.createInstance(
             SpreadsheetSkeleton,
             worksheet,
             config,
             worksheet.getCellMatrix(),
-            workbook.getStyles()
+            this._context.unit.getStyles()
         );
 
         return spreadsheetSkeleton;
